requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
bcrypt==4.1.2
pyjwt==2.8.0
